Refactor code out of application helper.
